Resolved Conflicts:
	README.md
	app/lib/enketo-transformer/enketo-xslt
	app/views/surveys/component/_enketo-power.jade
	config/default-config.json
	test/server/account-model.spec.js

+++ setup/bootstrap.sh
@@ -15,42 +15,32 @@ add-apt-repository -y ppa:rwky/redis
 apt-get update
 apt-get install -y redis-server
 
-# update repo
-echo 'updating enketo app to latest version'
-apt-get install -y git
-cd $ENKETO_EXPRESS_REPO_DIR
-# The next line should be commented out during development or any case where the repository was cloned via the ssh URL (i.e. git@github.com:kobotoolbox/enketo-express.git) as opposed to the HTTPS.
-[ $ENKETO_EXPRESS_UPDATE_REPO = "true" ] && git pull
-git submodule update --init --recursive
-
 # further redis setup with persistence, security, logging, multiple instances, priming 
-stop redis-server
 echo 'copying enketo redis conf...'
-mv /etc/redis/redis.conf redis-origin.conf
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/conf/redis-enketo-main.conf /etc/redis/
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/conf/redis-enketo-cache.conf /etc/redis/
-chown redis:redis /var/lib/redis/
-echo 'copying enketo redis-server configs...'
-mv /etc/init/redis-server.conf /etc/init/redis-server.conf.disabled
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/init/redis-server-enketo-main.conf /etc/init/
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/init/redis-server-enketo-cache.conf /etc/init/
-if [ -f "/var/lib/redis/redis.rdb" ]; then
-	rm /var/lib/redis/redis.rdb
+if [ -f "/etc/redis/redis.conf" ]; then
+    mv /etc/redis/redis.conf /etc/redis/redis-origin.conf
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/conf/redis-enketo-main.conf /etc/redis/
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/conf/redis-enketo-cache.conf /etc/redis/
+    chown redis:redis /var/lib/redis/
+    echo 'copying enketo redis-server configs...'
+    mv /etc/init/redis-server.conf /etc/init/redis-server.conf.disabled
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/init/redis-server-enketo-main.conf /etc/init/
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/init/redis-server-enketo-cache.conf /etc/init/
+    if [ -f "/var/lib/redis/redis.rdb" ]; then
+	   rm /var/lib/redis/redis.rdb
+    fi
+    echo 'copying enketo default redis db...'
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/enketo-main.rdb /var/lib/redis/
+    chown redis:redis /var/lib/redis/enketo-main.rdb
+    chmod 660 /var/lib/redis/enketo-main.rdb
 fi
-echo 'copying enketo default redis db...'
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/enketo-main.rdb /var/lib/redis/
-chown redis:redis /var/lib/redis/enketo-main.rdb
-chmod 660 /var/lib/redis/enketo-main.rdb
 echo 'starting first enketo redis instance...'
-service redis-server-enketo-main start
+service redis-server-enketo-main restart
 echo 'starting second enketo redis instance...'
-service redis-server-enketo-cache start
-
-# install XML prerequisites for node_xslt
-apt-get install -y libxml2-dev libxslt1-dev
+service redis-server-enketo-cache restart
 
 # install dependencies, development tools, node, grunt
-apt-get install -y python-software-properties python g++ make
+apt-get install -y python-software-properties python g++ make git
 cd $ENKETO_EXPRESS_REPO_DIR
 if [ $ENKETO_EXPRESS_USE_NODE_ENV = "true" ]; then
     apt-get install python-pip
@@ -58,33 +48,38 @@ if [ $ENKETO_EXPRESS_USE_NODE_ENV = "true" ]; then
     nodeenv env
     . env/bin/activate
 else
-    add-apt-repository ppa:chris-lea/node.js
-    apt-get update
+    curl -sL https://deb.nodesource.com/setup_0.10 | sudo bash -
     apt-get install -y nodejs
 fi
-npm install -g grunt-cli nodemon mocha bower
+npm -g install grunt-cli bower node-gyp gulp nodemon mocha
+npm -g install npm@next
 # remove node_modules if exists because npm builds can be system-specific
 if [ -d "$ENKETO_EXPRESS_REPO_DIR/node_modules" ]; then
 	rm -R $ENKETO_EXPRESS_REPO_DIR/node_modules
 fi
-npm install 
+npm install
 bower install --allow-root
 
-# build js and css
-grunt symlink
-grunt compile
+# create a local configuration file unless it already exists
+echo 'copying custom configuration unless config.json already exists'
+if [ ! -f "$ENKETO_EXPRESS_REPO_DIR/config/config.json" ]; then
+    cp setup/config/config.json config/config.json
+fi
 
-# start in background and keep restarting if it fails until `pm2 stop enketo` is called
-# developers may want to comment this out
+# build js and css
+grunt
+
+# still installing pm2 but not using in script any more. Because it runs invisibly when started as root.
 if [ $(whoami) = "root" ]; then
     npm install pm2@latest -g --unsafe-perm
 else
     npm install pm2@latest -g
 fi
-pm2 start app.js -n enketo
-echo "*************************************************************************************"
-echo "***                    Enketo Express should now have started!                   ****"
-echo "***                                                                              ****"
-echo "***   You can terminate it by ssh-ing into the VM and running: pm2 stop enketo   ****"
-echo "*************************************************************************************"
+
+echo "**************************************************************************************"
+echo "***                        Enketo Express is installed!                           ****"
+echo "***                                                                               ****"
+echo "*** You can start it by ssh-ing into the VM and running: cd /vagrant && npm start ****"
+echo "***                 ( or with: pm2 start /vagrant/app.js -n enketo )              ****"
+echo "**************************************************************************************"
